+++ app/src/main/java/com/example/facialexpressionchampionship/data/BattleHistoryRepository.kt
@@ -5,22 +5,12 @@ import com.example.facialexpressionchampionship.data.room.BattleInformationEntit
 import com.example.facialexpressionchampionship.data.room.ChallengerEntity
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Single
-import javax.inject.Inject
 
-class BattleHistoryRepository @Inject constructor(
-    private val localData: LocalData
-) : BattleHistorySource {
+interface BattleHistoryRepository {
 
-    override fun saveBattleInformation(battleInformation: BattleInformationEntity): Completable {
-        return localData.saveBattleInformation(battleInformation)
-    }
+    fun saveBattleInformation(battleInformation: BattleInformationEntity): Completable
 
-    override fun saveChallenger(challenger: List<ChallengerEntity>): Completable {
-        return localData.saveChallenger(challenger)
-    }
+    fun saveChallenger(challenger: List<ChallengerEntity>): Completable
 
-    override fun getBattleHistory(): Single<List<BattleHistory>> {
-        return localData.getBattleHistory()
-    }
-
+    fun getBattleHistory(): Single<List<BattleHistory>>
 }
