+++ app/src/main/java/com/example/facialexpressionchampionship/data/BattleHistoryRepositoryImpl.kt
@@ -1,0 +1,26 @@
+package com.example.facialexpressionchampionship.data
+
+import com.example.facialexpressionchampionship.data.room.BattleHistory
+import com.example.facialexpressionchampionship.data.room.BattleInformationEntity
+import com.example.facialexpressionchampionship.data.room.ChallengerEntity
+import io.reactivex.rxjava3.core.Completable
+import io.reactivex.rxjava3.core.Single
+import javax.inject.Inject
+
+class BattleHistoryRepositoryImpl @Inject constructor(
+    private val localData: LocalData
+) : BattleHistoryRepository {
+
+    override fun saveBattleInformation(battleInformation: BattleInformationEntity): Completable {
+        return localData.saveBattleInformation(battleInformation)
+    }
+
+    override fun saveChallenger(challenger: List<ChallengerEntity>): Completable {
+        return localData.saveChallenger(challenger)
+    }
+
+    override fun getBattleHistory(): Single<List<BattleHistory>> {
+        return localData.getBattleHistory()
+    }
+
+}
